Cleanup/Refactor shared heatmap code
