+++ keyboards/massdrop/ctrl/keymaps/beiufin/rules.mk
@@ -23,8 +23,8 @@ AUDIO_ENABLE = no           	# Audio output
 VIRTSER_ENABLE = no         	# USB Serial Driver
 RAW_ENABLE = no             	# Raw device
 AUTO_SHIFT_ENABLE = no      	# Auto Shift
-TAP_REPEAT_ENABLE = no          # Tap Repeat
-KEY_LOCK_TOGGLE_ENABLE = no     # Key Lock Toggle
+TAP_REPEAT_ENABLE = no      	# Tap Repeat
+KEY_LOCK_TOGGLE_ENABLE = no 	# Key Lock Toggle
 
 # Custom RGB matrix handling
 RGB_MATRIX_ENABLE = yes
